experiments/preferences/pyproject.toml
[project]
name = "identity-preferences"
version = "0.1.0"
description = "Experiment to discover which LLM personas prefer which other personas"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "polars>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
run-experiment = "scripts.run_experiment:app"
analyze-results = "scripts.analyze_results:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/persona_preferences"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
